+++ spec/outputs/tcp_spec.rb
@@ -108,6 +108,20 @@ describe LogStash::Outputs::Tcp do
     end
   end
 
+  context "with cipher suites" do
+    let(:config) do
+      super().merge('ssl_cipher_suites' => %w[TLS_RSA_WITH_AES_128_GCM_SHA256 TLS_DHE_RSA_WITH_3DES_EDE_CBC_SHA])
+    end
+
+    it "limits the ciphers selection" do
+      ssl_context = OpenSSL::SSL::SSLContext.new
+      allow(subject).to receive(:new_ssl_context).and_return(ssl_context)
+      subject.send :setup_ssl
+      expect(ssl_context.ciphers.length).to eq(2)
+      expect(ssl_context.ciphers).to satisfy { |arr| arr[0].include?('AES128-GCM-SHA256') && arr[1].include?('EDH-RSA-DES-CBC3-SHA') }
+    end
+  end
+
   context "client mode" do
     before { subject.register }
 
@@ -136,7 +150,7 @@ describe LogStash::Outputs::Tcp do
   end
 
   context "when enabling SSL" do
-    let(:config) { super().merge("ssl_enable" => true, 'codec' => 'plain') }
+    let(:config) { super().merge("ssl_enabled" => true, 'codec' => 'plain') }
     context "and not providing a certificate/key pair" do
       it "registers without error" do
         expect { subject.register }.to_not raise_error
@@ -145,13 +159,18 @@ describe LogStash::Outputs::Tcp do
 
     context "and providing a certificate/key pair" do
       let(:cert_key_pair) { Flores::PKI.generate }
-      let(:certificate) { cert_key_pair.first }
-      let(:cert_file) do
-        path = Tempfile.new('foo').path
-        IO.write(path, certificate.to_s)
+      let(:certificate) do
+        path = Tempfile.new('certificate').path
+        IO.write(path, cert_key_pair.first.to_s)
         path
       end
-      let(:config) { super().merge("ssl_cert" => cert_file) }
+      let(:key) do
+        path = Tempfile.new('key').path
+        IO.write(path, cert_key_pair[1].to_s)
+        path
+      end
+      let(:config) { super().merge("ssl_certificate" => certificate, "ssl_key" => key) }
+
       it "registers without error" do
         expect { subject.register }.to_not raise_error
       end
@@ -162,7 +181,7 @@ describe LogStash::Outputs::Tcp do
     context "ES generated plain-text certificate/key" do
       let(:key_file) { File.join(FIXTURES_PATH, 'plaintext/instance.key') }
       let(:crt_file) { File.join(FIXTURES_PATH, 'plaintext/instance.crt') }
-      let(:config) { super().merge("ssl_cert" => crt_file, "ssl_key" => key_file) }
+      let(:config) { super().merge("ssl_certificate" => crt_file, "ssl_key" => key_file) }
 
       it "registers without error" do
         expect { subject.register }.to_not raise_error
@@ -238,7 +257,7 @@ describe LogStash::Outputs::Tcp do
     context "encrypted key using PKCS#1" do
       let(:key_file) { File.join(FIXTURES_PATH, 'encrypted/instance.key') }
       let(:crt_file) { File.join(FIXTURES_PATH, 'encrypted/instance.crt') }
-      let(:config) { super().merge("ssl_cert" => crt_file, "ssl_key" => key_file) }
+      let(:config) { super().merge("ssl_certificate" => crt_file, "ssl_key" => key_file) }
 
       it "registers with error (due missing password)" do
         expect { subject.register }.to raise_error(OpenSSL::PKey::RSAError) # TODO need a better error
@@ -258,7 +277,7 @@ describe LogStash::Outputs::Tcp do
     context "and protocol is TLSv1.3" do
       let(:key_file) { File.join(FIXTURES_PATH, 'plaintext/instance.key') }
       let(:crt_file) { File.join(FIXTURES_PATH, 'plaintext/instance.crt') }
-      let(:config) { super().merge("ssl_cert" => crt_file, "ssl_key" => key_file) }
+      let(:config) { super().merge("ssl_certificate" => crt_file, "ssl_key" => key_file) }
 
       let(:secure_server) do
         ssl_context = OpenSSL::SSL::SSLContext.new
@@ -305,6 +324,145 @@ describe LogStash::Outputs::Tcp do
         thread.join(2)
 
         expect(buffer).to eq(message * 2)
+      end
+    end
+
+    context "with only ssl_certificate set" do
+      let(:config) { super().merge("ssl_certificate" => File.join(FIXTURES_PATH, 'plaintext/instance.crt')) }
+
+      it "should raise a configuration error to request also `ssl_key`" do
+        expect { subject.register }.to raise_error(LogStash::ConfigurationError, /Using an `ssl_certificate` requires an `ssl_key`/)
+      end
+    end
+
+    context "with only ssl_key set" do
+      let(:config) { super().merge("ssl_key" => File.join(FIXTURES_PATH, 'plaintext/instance.key')) }
+
+      it "should raise a configuration error to request also `ssl_key`" do
+        expect { subject.register }.to raise_error(LogStash::ConfigurationError, /An `ssl_certificate` is required when using an `ssl_key`/)
+      end
+    end
+
+    context "and mode is server" do
+      let(:config) do
+        {
+          "host" => "127.0.0.1",
+          "port" => port,
+          "mode" => 'server',
+          "ssl_enabled" => true,
+          "ssl_certificate" => File.join(FIXTURES_PATH, 'plaintext/instance.crt'),
+          "ssl_key" => File.join(FIXTURES_PATH, 'plaintext/instance.key'),
+        }
+      end
+
+      context "with no ssl_certificate" do
+        let(:config) { super().reject { |k| "ssl_key".eql?(k) || "ssl_certificate".eql?(k) } }
+
+        it "should raise a configuration error" do
+          expect { subject.register }.to raise_error(LogStash::ConfigurationError, /An `ssl_certificate` is required when `ssl_enabled` => true/)
+        end
+      end
+
+      context "with ssl_client_authentication = `none` and no ssl_certificate_authorities" do
+        let(:config) { super().merge(
+          'ssl_client_authentication' => 'none',
+          'ssl_certificate_authorities' => []
+        ) }
+
+        it "should register without errors" do
+          expect { subject.register }.to_not raise_error
+        end
+      end
+
+      context "with deprecated ssl_verify = true and no ssl_certificate_authorities" do
+        let(:config) { super().merge(
+          'ssl_verify' => true,
+          'ssl_certificate_authorities' => []
+        ) }
+
+        it "should register without errors" do
+          expect { subject.register }.to_not raise_error
+        end
+      end
+
+      %w[required optional].each do |ssl_client_authentication|
+        context "with ssl_client_authentication = `#{ssl_client_authentication}` and no ssl_certificate_authorities" do
+          let(:config) { super().merge(
+            'ssl_client_authentication' => ssl_client_authentication,
+            'ssl_certificate_authorities' => []
+          ) }
+
+          it "should raise a configuration error" do
+            expect { subject.register }.to raise_error(LogStash::ConfigurationError, /An `ssl_certificate_authorities` is required when `ssl_client_authentication` => `#{ssl_client_authentication}`/)
+          end
+        end
+      end
+
+      context "with ssl_verification_mode" do
+        let(:config) do
+          super().merge 'ssl_verification_mode' => 'full'
+        end
+
+        it "should raise a configuration error" do
+          expect{subject.register}.to raise_error(LogStash::ConfigurationError, /`ssl_verification_mode` must not be configured when mode is `server`, use `ssl_client_authentication` instead/)
+        end
+      end
+    end
+
+    context "with deprecated settings" do
+      let(:ssl_verify) { true }
+      let(:certificate_path) { File.join(FIXTURES_PATH, 'plaintext/instance.crt') }
+      let(:config) do
+        {
+          "host" => "127.0.0.1",
+          "port" => port,
+          "ssl_enable" => true,
+          "ssl_cert" => certificate_path,
+          "ssl_key" => File.join(FIXTURES_PATH, 'plaintext/instance.key'),
+          "ssl_verify" => ssl_verify
+        }
+      end
+
+      context "and mode is server" do
+        let(:config) { super().merge("mode" => 'server') }
+        [true, false].each do |verify|
+          context "and ssl_verify is #{verify}" do
+            let(:ssl_verify) { verify }
+
+            it "should set new configs variables" do
+              subject.register
+              expect(subject.instance_variable_get(:@ssl_enabled)).to eql(true)
+              expect(subject.instance_variable_get(:@ssl_client_authentication)).to eql(verify ? 'required' : 'none')
+              expect(subject.instance_variable_get(:@ssl_certificate)).to eql(certificate_path)
+            end
+          end
+        end
+      end
+
+      context "and mode is client" do
+        let(:config) { super().merge("mode" => 'client') }
+        [true, false].each do |verify|
+          context "and ssl_verify is #{verify}" do
+            let(:ssl_verify) { verify }
+
+            it "should set new configs variables" do
+              subject.register
+              expect(subject.instance_variable_get(:@ssl_enabled)).to eql(true)
+              expect(subject.instance_variable_get(:@ssl_verification_mode)).to eql(verify ? 'full' : 'none')
+              expect(subject.instance_variable_get(:@ssl_certificate)).to eql(certificate_path)
+            end
+          end
+        end
+      end
+    end
+
+    context "with ssl_client_authentication" do
+      let(:config) do
+        super().merge 'ssl_client_authentication' => 'required'
+      end
+
+      it "should raise a configuration error" do
+        expect{subject.register}.to raise_error(LogStash::ConfigurationError, /`ssl_client_authentication` must not be configured when mode is `client`, use `ssl_verification_mode` instead/)
       end
     end
   end